move labels into own module
